requirements.txt
alembic==1.4.2
altgraph==0.17
asgiref==3.2.3
astroid==2.4.1
asymcrypt==0.0.10
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.2
base32hex==1.0.2
beautifulsoup4==4.8.2
blinker==1.4
bs4==0.0.1
camelot-py==0.7.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
click-default-group==1.2.2
cloudinary==1.20.0
colorama==0.4.3
colouration==2020.4.28
constantly==15.1.0
cryptography==2.9
cssselect==1.1.0
cycler==0.10.0
db-to-sqlite==1.1.1
defusedxml==0.6.0
dill==0.3.1.1
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.2
django-crispy-forms==1.9.1
django-filters==0.2.1
django-heroku==0.3.1
docutils==0.16
EasyProcess==0.3
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.8.2
flake8-docstrings==1.5.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Jsonpify==1.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Mobility==1.0.0
Flask-MongoAlchemy==0.7.2
Flask-MySQL==1.5.1
Flask-Pusher==3.0
Flask-PyMongo==2.3.0
Flask-QRcode==3.0.0
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
func-timeout==4.3.5
future==0.18.2
geographiclib==1.50
geojson==2.5.0
geopy==1.21.0
gTTS==2.0.4
gTTS-token==1.1.3
gunicorn==20.0.4
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
inflect==4.1.0
isort==4.3.21
itsdangerous==1.1.0
jaraco.itertools==5.0.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.1
json5==0.9.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.4.2
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mechanize==0.4.5
MongoAlchemy==0.19
more-itertools==8.2.0
ndg-httpsclient==0.5.1
numpy==1.18.0
opencv-python==4.2.0.32
openpyxl==3.0.3
pandas==0.25.3
parsel==1.5.2
parso==0.7.0
passlib==1.7.2
pbr==5.4.4
pdfminer.six==20200124
pefile==2019.4.18
Pillow==7.1.1
pinterest-api==0.0.8
playsound==1.2.2
pluggy==0.13.1
Protego==0.1.16
psycopg2==2.8.5
pusher==2.1.4
py-crypt==1.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.6
pycurl==7.43.0.5
PyDispatcher==2.0.5
pydocstyle==5.0.2
pyfiglet==0.8.post1
pyflakes==2.2.0
pygame==1.9.6
Pygments==2.5.2
PyHamcrest==2.0.2
PyInstaller==3.6
pylint==2.5.2
pymongo==2.8.1
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyowm==3.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pypiwin32==223
PyQt5==5.13.2
PyQt5-sip==12.7.1
pyqt5-tools==5.13.2.1.6rc1
pyquery==1.4.1
PySocks==1.7.1
pyspider==0.3.10
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.12.0
python-editor==1.0.4
python-engineio==3.12.1
python-jsonrpc-server==0.3.4
python-language-server==0.33.0
python-socketio==4.5.1
pyttsx3==2.71
pytz==2019.3
PyVirtualDisplay==0.2.5
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.3.1
qrcode==6.1
requests==2.22.0
robobrowser==0.5.3
rope==0.17.0
rsa==4.0
scapy==2.4.3
selenium==3.141.0
service-identity==18.1.0
six==1.13.0
slytherin==2020.4.28
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
SQLAlchemy==1.3.15
sqlite-utils==2.3.1
sqlparse==0.3.0
stevedore==1.31.0
stripe==2.48.0
tabula-py==2.0.4
tabulate==0.8.6
tblib==1.6.0
toml==0.10.1
tornado==4.5.3
u-msgpack-python==2.5.2
urllib2-file==0.2.1
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
w3lib==1.21.0
webencodings==0.5.1
Werkzeug==1.0.0
whitenoise==5.1.0
wolframalpha==4.0.0
wrapt==1.12.1
WsgiDAV==3.0.3
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.2.7
xmltodict==0.12.0
yapf==0.30.0
zope.interface==5.1.0
